Created rgrep.c
rgrep is a variant of the grep function, where it takes a search term (known as pattern) and searches throughout user given inputs or a plain text file. 
It will match base cases (regular words) alongside some symbols that the user may have in their search term.

The special characters (or wildcard characters are as follows):
+ . (period) - Matches any character
+ + (plus) - The preceding character will appear one or more times
+ ? (question mark) - The preceding character may or may not appear.
+ \ (backslash) - "Escapes" the following character, nullifying any special meaning it may have

The code in here right now is just the "skeleton" included in the project files.